requirements.txt
streamlit
pandas
numpy
networkx
pillow
matplotlib
plotly